--- Project-1.Data/Location.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Project_1.Data
-{
-    public class Location
-    {
-        public int X { get; }
-
-        public int Y { get; }
-
-        public Location(int x, int y)
-        {
-            X = x;
-            Y = y;
-        }
-    }
-}
